treewide: re-export patches autonominated for refresh [4.6 --> 4.6.1]
When creating a tree from scratch with the yocto-kernel-tools, the
commit application is first attempted with "git am".  If that fails
it then falls back to the slightly less strict default arguments of
"patch" and constructs the rest of the commit using that.  At the
same time, it creates an entry for the file that failed and the
commit ID of the reconstructed commit.

Here we "git format-patch" out the commits into their respective
files in order to refresh them so they can be git-am'd and not
use the fallback method.  This does not change the source code of
the generated tree, as easily verified by "git diff" and confirming
the pre and post trees on either side of this change are equivalent.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>